requirements.txt
fastapi
uvicorn
python-multipart
streamlit
requests
opencv-python
numpy
ultralytics
pillow
